account: changed float to BigDecimal

+++ src/main/java/com/nlogax/banking/dto/TransactionDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,5 +13,5 @@ public class TransactionDto {
     private String id;
     private String numberFrom;
     private String numberTo;
-    private float amount;
+    private BigDecimal amount;
 }
